demo-api-main/requirements.txt
# Web Framework
Flask==3.0.0
flask-restful==0.3.9
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1

# Data Processing
pandas==2.1.4

# Google Cloud Libraries
google-cloud-dlp>=3.0
google-cloud-kms>=2.0
gunicorn>=20.0 # For Cloud Run deployment

# Utilities
PyYAML==6.0.1
bcrypt==4.0.1